llm-kernel/pyproject.toml
[project]
name = "llm-kernel"
version = "0.1.0"
description = "LLM Kernel for Chatroom.AI by Unchpokable"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai",
    "httpx",
    "python-dotenv",
    "tiktoken",
    "pydantic",
    "uvicorn[standard]>=0.40.0",
    "fastapi>=0.128.0",
]

[dependency-groups]
dev = [
    "grpcio-tools>=1.76.0",
]
